core/src/test/java/com/avolution/net/tcp/TCPClientServiceTest.java
package com.avolution.net.tcp;

import com.avolution.net.MessagePacket;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.util.concurrent.TimeUnit;

import static org.junit.jupiter.api.Assertions.*;

class TCPClientServiceTest {

    private TCPClientService tcpClientService;

    @BeforeEach
    void setUp() {
        tcpClientService = new TCPClientService("127.0.0.1", 8080);
    }

    @Test
    void testSendString() {
        String content = "Hello, Server!";
        tcpClientService.send(content);
        // Add assertions to verify the string was sent correctly
    }

    @Test
    void testSendByteArray() {
        byte[] content = "Hello, Server!".getBytes();
        tcpClientService.send(content);
        // Add assertions to verify the byte array was sent correctly
    }

    @Test
    void testConnectionPooling() {
        // Add test logic to verify connection pooling functionality
    }

    @Test
    void testConnectionTimeoutHandling() {
        // Add test logic to verify connection timeout handling functionality
    }

    @Test
    void testAsynchronousIO() {
        // Add test logic to verify asynchronous I/O operations
    }

    @Test
    void testMessagePacketFunctionality() {
        byte[] content = "Test Message".getBytes();
        MessagePacket packet = new MessagePacket(content.length, content);
        assertEquals(content.length, packet.getLength());
        assertArrayEquals(content, packet.getContent());
    }

    @Test
    void testStartService() throws InterruptedException {
        tcpClientService.start();
        assertTrue(tcpClientService.isRunning());
    }

    @Test
    void testPauseService() {
        tcpClientService.pause();
        assertEquals(IService.Status.PAUSED, tcpClientService.getStatus());
    }

    @Test
    void testStopService() {
        tcpClientService.stop();
        assertEquals(IService.Status.STOPPED, tcpClientService.getStatus());
    }

    @Test
    void testRestartService() throws InterruptedException {
        tcpClientService.restart();
        assertTrue(tcpClientService.isRunning());
    }

    @Test
    void testSessionManagement() {
        // Add test logic to verify session management functionality
    }
}
